rename two files
