+++ jsapp/js/components/common/IconLegacySvgMappings.tsx
